Reduce duplicate code of importing test cases
Importing test cases to a plan via both Web UI and XMLRPC has almost
the same code. This patch aims to reduce the duplicate and share same
code.

TestPlan has a new interface import_cases to import cases loaded from
XML.

Fixes #280
