+++ tcms/xmlrpc/tests/test_testplan.py
@@ -1,43 +1,34 @@
 # -*- coding: utf-8 -*-
 
 import unittest
-import xmltodict
 
 from six.moves.http_client import BAD_REQUEST
 from six.moves.http_client import NOT_FOUND
 from six.moves.http_client import INTERNAL_SERVER_ERROR
 
 from django import test
-from django.conf import settings
-from django.contrib.auth.models import User
 
-from tcms.management.models import Priority
-from tcms.management.models import TestTag
 from tcms.testcases.models import TestCase
 from tcms.testcases.models import TestCasePlan
-from tcms.testcases.models import TestCaseStatus
-from tcms.testplans.models import TestPlan
 from tcms.testplans.models import TCMSEnvPlanMap
-from tcms.xmlrpc.api import testplan as XmlrpcTestPlan
-from tcms.xmlrpc.api.testplan import clean_xml_file
-from tcms.xmlrpc.api.testplan import process_case
-from tcms.xmlrpc.tests.utils import make_http_request
-
+from tcms.testplans.models import TestPlan
 from tcms.tests.factories import ComponentFactory
 from tcms.tests.factories import ProductFactory
+from tcms.tests.factories import TCMSEnvGroupFactory
 from tcms.tests.factories import TestCaseFactory
 from tcms.tests.factories import TestPlanFactory
 from tcms.tests.factories import TestPlanTypeFactory
 from tcms.tests.factories import TestTagFactory
-from tcms.tests.factories import TCMSEnvGroupFactory
 from tcms.tests.factories import UserFactory
 from tcms.tests.factories import VersionFactory
+from tcms.xmlrpc.api import testplan as XmlrpcTestPlan
+from tcms.xmlrpc.api.testplan import import_case_via_XML
 from tcms.xmlrpc.tests.utils import XmlrpcAPIBaseTest
+from tcms.xmlrpc.tests.utils import make_http_request
 
 __all__ = (
     'TestAddComponent',
     'TestAddTag',
-    'TestCleanXMLFile',
     'TestComponentMethods',
     'TestFilter',
     'TestGetAllCasesTags',
@@ -311,11 +302,6 @@ class TestRemoveTag(test.TestCase):
     """TODO: """
 
 
-@unittest.skip('TODO: test case is not implemented yet.')
-class TestImportCaseViaXML(test.TestCase):
-    """TODO: """
-
-
 class TestUpdate(test.TestCase):
     """ Tests the XMLRPM testplan.update method """
 
@@ -363,273 +349,21 @@ class TestUpdate(test.TestCase):
         self.assertEqual(2, TCMSEnvPlanMap.objects.filter(plan__in=[self.plan_1, self.plan_2]).count())
 
 
-# ################ Section for testing testplan.import_case_via_XML ########
-
-xml_single_case = '''
-<testcase author="%(author)s" priority="%(priority)s"
-          automated="%(automated)s" status="%(status)s">
-    <summary>%(summary)s</summary>
-    <categoryname>%(categoryname)s</categoryname>
-    <defaulttester>%(defaulttester)s</defaulttester>
-    <notes>%(notes)s</notes>
-    <action>%(action)s</action>
-    <expectedresults>%(expectedresults)s</expectedresults>
-    <setup>%(setup)s</setup>
-    <breakdown>%(breakdown)s</breakdown>
-    %(tags)s
-</testcase>'''
-
-
-sample_case_data = {
-    'author': 'user@example.com',
-    'priority': 'P1',
-    'automated': '0',
-    'status': 'CONFIRMED',
-    'summary': 'test case',
-    'categoryname': '--default--',
-    'defaulttester': '',
-    'notes': '',
-    'action': '',
-    'expectedresults': '',
-    'setup': '',
-    'effect': '',
-    'breakdown': '',
-    'tag': ['tag 1'],
-}
-
-
-xml_file_without_error = u'''
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<!DOCTYPE testopia SYSTEM "testopia.dtd" [
-  <!ENTITY testopia_lt "<">
-  <!ENTITY testopia_gt ">">
-]>
-<testopia version="1.1">
-  <testcase author="user@example.com" priority="P1"
-      automated="0" status="CONFIRMED">
-    <summary>测试用例</summary>
-    <categoryname>--default--</categoryname>
-    <defaulttester></defaulttester>
-    <notes>&lt;script type=&quot;text/javascript&quot;&gt;
-    alert(&quot;Exploited!&quot;);
-    &lt;/script&gt;</notes>
-    <action></action>
-    <expectedresults></expectedresults>
-    <setup></setup>
-    <breakdown></breakdown>
-    <tag>haha &lt;script&gt;alert(&#39;HAHAHA&#39;)&lt;/script&gt;</tag>
-  </testcase>
-  <testcase author="user@example.com" priority="P1"
-      automated="0" status="CONFIRMED">
-    <summary>case 2</summary>
-    <categoryname>--default--</categoryname>
-    <defaulttester></defaulttester>
-    <notes></notes>
-    <action></action>
-    <expectedresults></expectedresults>
-    <setup></setup>
-    <breakdown></breakdown>
-    <tag>xmlrpc</tag>
-    <tag>haha</tag>
-    <tag>case management system</tag>
-  </testcase>
-</testopia>
-'''
-
-
-xml_file_single_case_without_error = u'''
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<!DOCTYPE testopia SYSTEM "testopia.dtd" [
-  <!ENTITY testopia_lt "<">
-  <!ENTITY testopia_gt ">">
-]>
-<testopia version="1.1">
-  <testcase author="user@example.com" priority="P1"
-      automated="0" status="CONFIRMED">
-    <summary>case 1</summary>
-    <categoryname>--default--</categoryname>
-    <defaulttester></defaulttester>
-    <notes>&lt;script type=&quot;text/javascript&quot;&gt;
-    alert(&quot;Exploited!&quot;);
-    &lt;/script&gt;</notes>
-    <action></action>
-    <expectedresults></expectedresults>
-    <setup></setup>
-    <breakdown></breakdown>
-    <tag>haha &lt;script&gt;alert(&#39;HAHAHA&#39;)&lt;/script&gt;</tag>
-  </testcase>
-</testopia>
-'''
-
-
-# With error, non-existent priority and defaulttester's email
-xml_file_with_error = u'''
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<!DOCTYPE testopia SYSTEM "testopia.dtd" [
-  <!ENTITY testopia_lt "<">
-  <!ENTITY testopia_gt ">">
-]>
-<testopia version="1.1">
-  <testcase author="user@example.com" priority="Pn"
-      automated="0" status="CONFIRMED">
-    <summary>case 2</summary>
-    <categoryname>--default--</categoryname>
-    <defaulttester>x-man@universe.net</defaulttester>
-    <notes></notes>
-    <action></action>
-    <expectedresults></expectedresults>
-    <setup></setup>
-    <breakdown></breakdown>
-    <tag>xmlrpc</tag>
-    <tag>haha</tag>
-    <tag>case management system</tag>
-  </testcase>
-</testopia>
-'''
-
-
-xml_file_with_wrong_version = u'''
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<!DOCTYPE testopia SYSTEM "testopia.dtd" [
-  <!ENTITY testopia_lt "<">
-  <!ENTITY testopia_gt ">">
-]>
-<testopia version="who knows"></testopia>'''
-
-
-xml_file_in_malformat = u'''
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<!DOCTYPE testopia SYSTEM "testopia.dtd" [
-  <!ENTITY testopia_lt "<">
-  <!ENTITY testopia_gt ">">
-]>
-<nitrate version="1.1"></nitrate>'''
-
-
-class TestProcessCase(test.TestCase):
-    """Test process_case"""
+class TestImportCaseViaXML(XmlrpcAPIBaseTest):
+    """Test import_case_via_XML"""
 
     @classmethod
     def setUpTestData(cls):
-        cls.user = UserFactory(username='xml user', email='user@example.com')
-        cls.priority_p1, created = Priority.objects.get_or_create(value='P1')
-        cls.status_confirmed, created = TestCaseStatus.objects.get_or_create(name='CONFIRMED')
-        cls.status_proposed, created = TestCaseStatus.objects.get_or_create(name='PROPOSED')
+        cls.author = UserFactory(username='user', email='user@example.com')
+        cls.plan = TestPlanFactory(author=cls.author, owner=cls.author)
+        cls.http_req = make_http_request(user=cls.author)
 
-    def _format_xml_case_string(self, case_data):
-        """Helper method to format case XML conveniently"""
-        data = case_data.copy()
-        data['tags'] = ''.join(
-            '<tag>{0}</tag>'.format(tag) for tag in data['tag'])
-        data.pop('tag')
-        return xml_single_case % data
+    def test_succeed_to_import_cases(self):
+        from tcms.testplans.tests.test_importer import xml_file_without_error
+        result = import_case_via_XML(self.http_req,
+                                     self.plan.pk,
+                                     xml_file_without_error)
 
-    def _create_xml_dict(self, case_data):
-        xml_case = self._format_xml_case_string(case_data)
-        return xmltodict.parse(xml_case)
-
-    def test_process_case(self):
-        xmldict = self._create_xml_dict(sample_case_data)
-
-        cleaned_case = process_case(xmldict['testcase'])
-        self.assertEqual(self.user.id, cleaned_case['author_id'])
-        self.assertEqual(self.user, cleaned_case['author'])
-        self.assertEqual(sample_case_data['summary'], cleaned_case['summary'])
-        self.assertEqual(None, cleaned_case['default_tester_id'])
-        p1 = Priority.objects.get(value=sample_case_data['priority'])
-        self.assertEqual(p1.pk, cleaned_case['priority_id'])
-        self.assertEqual(False, cleaned_case['is_automated'])
-        self.assertEqual(sample_case_data['categoryname'],
-                         cleaned_case['category_name'])
-        self.assert_(isinstance(cleaned_case['tags'], list))
-        for tag in sample_case_data['tag']:
-            expected_tag = TestTag.objects.get(name=tag)
-            self.assertEqual(expected_tag, cleaned_case['tags'][0])
-        self.assertEqual(sample_case_data['action'], cleaned_case['action'])
-        self.assertEqual(sample_case_data['effect'], cleaned_case['effect'])
-        self.assertEqual(sample_case_data['setup'], cleaned_case['setup'])
-        self.assertEqual(sample_case_data['breakdown'],
-                         cleaned_case['breakdown'])
-        self.assertEqual(sample_case_data['notes'], cleaned_case['notes'])
-
-    def test_nonexistent_object(self):
-        case_data = sample_case_data.copy()
-        case_data['author'] = 'another_user@example.com'
-        xmldict = self._create_xml_dict(case_data)
-        self.assertRaises(User.DoesNotExist, process_case, xmldict['testcase'])
-
-        case_data = sample_case_data.copy()
-        case_data['defaulttester'] = 'another_user@example.com'
-        xmldict = self._create_xml_dict(case_data)
-        self.assertRaises(User.DoesNotExist, process_case, xmldict['testcase'])
-
-        case_data = sample_case_data.copy()
-        case_data['priority'] = 'PP'
-        xmldict = self._create_xml_dict(case_data)
-        self.assertRaises(Priority.DoesNotExist,
-                          process_case, xmldict['testcase'])
-
-        case_data = sample_case_data.copy()
-        case_data['priority'] = ''
-        xmldict = self._create_xml_dict(case_data)
-        self.assertRaises(ValueError, process_case, xmldict['testcase'])
-
-        case_data = sample_case_data.copy()
-        case_data['status'] = 'UNKNOWN_STATUS'
-        xmldict = self._create_xml_dict(case_data)
-        self.assertRaises(TestCaseStatus.DoesNotExist,
-                          process_case, xmldict['testcase'])
-
-        case_data = sample_case_data.copy()
-        case_data['status'] = ''
-        xmldict = self._create_xml_dict(case_data)
-        self.assertRaises(ValueError, process_case, xmldict['testcase'])
-
-        case_data = sample_case_data.copy()
-        case_data['automated'] = ''
-        xmldict = self._create_xml_dict(case_data)
-        cleaned_case = process_case(xmldict['testcase'])
-        self.assertEqual(False, cleaned_case['is_automated'])
-
-        case_data = sample_case_data.copy()
-        case_data['tag'] = ''
-        xmldict = self._create_xml_dict(case_data)
-        cleaned_case = process_case(xmldict['testcase'])
-        self.assertEqual(None, cleaned_case['tags'])
-
-        case_data = sample_case_data.copy()
-        case_data['categoryname'] = ''
-        xmldict = self._create_xml_dict(case_data)
-        self.assertRaises(ValueError, process_case, xmldict['testcase'])
-
-
-class TestCleanXMLFile(test.TestCase):
-    """Test for testplan.clean_xml_file"""
-
-    @classmethod
-    def setUpTestData(cls):
-        cls.user = UserFactory(username='xml user', email='user@example.com')
-        cls.priority = Priority.objects.get(value='P1')
-        cls.status_confirmed = TestCaseStatus.objects.get(name='CONFIRMED')
-
-        cls.original_xml_version = None
-        if hasattr(settings, 'TESTOPIA_XML_VERSION'):
-            cls.original_xml_version = settings.TESTOPIA_XML_VERSION
-        settings.TESTOPIA_XML_VERSION = '1.1'
-
-    def test_clean_xml_file(self):
-        result = clean_xml_file(xml_file_without_error)
-        self.assertEqual(2, len(list(result)))
-
-        result = clean_xml_file(xml_file_single_case_without_error)
-        self.assertEqual(1, len(list(result)))
-
-        cases = clean_xml_file(xml_file_with_error)
-        self.assertRaises(User.DoesNotExist, list, cases)
-
-        self.assertRaises(ValueError, clean_xml_file, xml_file_in_malformat)
-        self.assertRaises(
-            ValueError, clean_xml_file, xml_file_with_wrong_version)
-
-
-# ################ End of testing testplan.import_case_via_XML #############
+        self.assertEqual(2, TestCase.objects.count())
+        self.assertTrue(TestCase.objects.filter(summary='case 2').exists())
+        self.assertEqual('Success update 2 cases', result)
